Fixed long parsing in settings

+++ src/main/java/com/greendelta/collaboration/model/settings/Setting.java
@@ -50,7 +50,7 @@ public class Setting extends AbstractEntity {
 		if (type == Integer.class && value != null)
 			return (V) Integer.valueOf(Integer.parseInt(value));
 		if (type == Long.class && value != null)
-			return (V) Integer.valueOf(Integer.parseInt(value));
+			return (V) Long.valueOf(Long.parseLong(value));
 		if (type == Object.class && data != null) {
 			try {
 				String json = new String(data, StandardCharsets.UTF_8);
